modularize hydra command subprocess and move cli error catch to utils
update imports
