+++ autrainer/core/scripts/postprocess_script.py
@@ -8,7 +8,8 @@ from .abstract_postprocess_script import (
     AbstractPostprocessArgs,
     AbstractPostprocessScript,
 )
-from .command_line_error import CommandLineError, catch_cli_errors
+from .command_line_error import CommandLineError
+from .utils import catch_cli_errors
 
 
 @dataclass
