+++ autrainer/core/scripts/show_script.py
@@ -8,7 +8,8 @@ import autrainer
 from autrainer.core.constants import CONFIG_FOLDERS
 
 from .abstract_script import AbstractScript, MockParser
-from .command_line_error import CommandLineError, catch_cli_errors
+from .command_line_error import CommandLineError
+from .utils import catch_cli_errors
 
 
 @dataclass
